Ändringar inför inlämning Delat upp kod i olika filer för lättare överblick. Fixat den sista buggen med null-värden i listan. Mindre justeringar.
